AHCI: uint32_t vs. size_t issues, hopefully fixes windows buildboxes
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@64022 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Storage/DevAHCI.cpp
@@ -3927,7 +3927,7 @@ static int atapiPassthroughSS(PAHCIREQ pAhciReq, PAHCIPort pAhciPort, size_t cbD
     uint32_t cbTransfer;
     void *pvBuf = NULL;
 
-    cbTransfer = pAhciReq->cbTransfer;
+    cbTransfer = (uint32_t)pAhciReq->cbTransfer;
 
     if (cbTransfer)
     {
